fix: drops the country/state reference from the label
this is less than ideal, but for the short term this is the route I'm taking. The issue is the tspan appears to get removed on update since it is nested within a text element.

I will like to go back to his problem and address it at a different time.

fixes #67